new game and continue game work
