--- src/util.ts
@@ -1,4 +1,0 @@
-
-function hasSetUsername():boolean {
-    return document.cookie.split(';').some((item) => item.trim().startsWith('username='))
-}
